feat: :sparkles: set up extension (#2)
# Description

This PR introduces the Quarto extension for the ON-LiMiT theme,
including configuration updates, a new example page, and improved
documentation. It adds support for emojis, a favicon, and a consistent
navbar, while cleaning up unnecessary files.

Unfortunately, there's not an extension type specifically for brands
currently (but it's
[planned](https://github.com/quarto-dev/quarto-cli/discussions/12688)
for Quarto 1.8), so I used the [Custom
formats](https://quarto.org/docs/extensions/formats.html) extension type
for now.

I've tried to keep this extension as simple as possible 🍎

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>